# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Swagger",
    "title": "Recipes API Swagger",
    "contact": {},
    "version": "1.0"
  },
  "host": "localhost:8080",
  "paths": {
...
    "/pictures/{id}": {
      "get": {
        "description": "Get a profile picture",
        "produces": [
          "image/jpeg",
          "image/png"
        ],
        "tags": [
          "Users"
        ],
        "summary": "Get profile picture",
        "parameters": [
          {
            "type": "string",
            "example": "6567bd0a84f663dbe91176f5.png",
            "description": "Id of the picture with the extension",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Jpeg, png or jpg image"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Get users with filters",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Users"
        ],
        "summary": "Get users",
        "parameters": [
          {
            "type": "integer",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "name": "offset",
            "in": "query"
          },
          {
            "type": "string",
            "name": "username",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.GetUsersResponse"
            }
          },
          "400": {
            "description": "Couldn't read query parameters",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "401": {
            "description": "Invalid or expired token",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Failed to get users",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      }
    },
...
    "/users/{id}": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Get another user by its id",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Users"
        ],
        "summary": "Get user by id",
        "parameters": [
          {
            "type": "string",
            "description": "User id to get",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.UserView"
            }
          },
          "401": {
            "description": "Invalid or expired token",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "404": {
            "description": "User not found",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Delete a user. Must be admin",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Users"
        ],
        "summary": "Delete user",
        "parameters": [
          {
            "type": "string",
            "example": "6567bd0a84f663dbe91176f5",
            "description": "User id to delete",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted",
            "schema": {
              "$ref": "#/definitions/models.UserDB"
            }
          },
          "401": {
            "description": "Invalid or expired token",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "403": {
            "description": "Unknown / Insufficient permissions",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "404": {
            "description": "User not found",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Failed to delete the user / delete associated picture / delete parameters",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      },
      "patch": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Update another user by its id. Must be admin",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Users"
        ],
        "summary": "Update user",
        "parameters": [
          {
            "type": "string",
            "example": "6567bd0a84f663dbe91176f5",
            "description": "User id to update",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Infos to update",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.UserDB"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.UserDB"
            }
          },
          "400": {
            "description": "Bad request / Username too long",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "401": {
            "description": "Invalid or expired token",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "403": {
            "description": "Unknown / Insufficient permissions",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "404": {
            "description": "User not found",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "409": {
            "description": "Username or email already taken",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      }
    },
    "/users/{id}/picture": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Update a user's profile picture, must be under 8MB. Must be admin",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Users"
        ],
        "summary": "Update user picture",
        "parameters": [
          {
            "type": "string",
            "example": "6567bd0a84f663dbe91176f5",
            "description": "User id to update the picture",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "file",
            "description": "Picture, formData must be named 'file'",
            "name": "request",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.UpdatePictureResponse"
            }
          },
          "400": {
            "description": "Empty formData / not named correctly",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "401": {
            "description": "Invalid or expired token",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "403": {
            "description": "Unknown / Insufficient permissions",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "406": {
            "description": "Wrong format (not png / jpeg / jpg) / File too heavy",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Failed to: Update user, save the image",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Delete a user's profile picture. Must be admin",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Users"
        ],
        "summary": "Delete user picture",
        "parameters": [
          {
            "type": "string",
            "example": "6567bd0a84f663dbe91176f5",
            "description": "User id to delete the picture from",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Picture deleted",
            "schema": {
              "$ref": "#/definitions/models.MessageResponse"
            }
          },
          "401": {
            "description": "Invalid or expired token",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "403": {
            "description": "Unknown / Insufficient permissions",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "404": {
            "description": "Image not found",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "422": {
            "description": "No picture for this user",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Failed to: Update user / remove the picture",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "models.ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "code": {
          "type": "string",
          "description": "Stable machine-readable error code"
        },
        "request_id": {
          "type": "string",
          "description": "Request identifier, also returned in X-Request-ID"
        }
      }
    },
...
    "models.GetUsersResponse": {
      "type": "object",
      "properties": {
        "items": {},
        "length": {
          "type": "integer"
        }
      }
    },
...
    "models.MessageResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
...
    "models.UpdatePictureResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
...
    "models.UserDB": {
      "type": "object",
      "properties": {
        "admin": {
          "type": "boolean"
        },
        "email": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "picture": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    },
...
    "models.UserView": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "picture": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "basePath": "/api/v1"
}